made sizes 64 bit

+++ utils.c
@@ -159,7 +159,7 @@ FILE *fopen64_and_check(const char *fname, const char *mode, int error)
     return buffer;
 }
 
-void *calloc_and_check(size_t num, size_t size, char *msg)
+void *calloc_and_check(int64_t num, int64_t size, char *msg)
 {
     void *block;
     if ((block=calloc(num,size))==NULL)
